Split attribute list and use hashset!
